--- a/R/utils.R
+++ b/R/utils.R
@@ -182,7 +182,11 @@ list_of_ptype <- function(x) {
 }
 
 apply_names_sep <- function(outer, inner, names_sep) {
-  if (length(inner) == 0 || length(outer) == 0) {
+  # Need to avoid `paste0()` recycling issue. Not using `vec_paste0()`
+  # because that is too slow to be applied to each element (#1427).
+  # `outer` and `names_sep` are required to be length 1,
+  # so we only need to check `inner`.
+  if (length(inner) == 0L) {
     character()
   } else {
     paste0(outer, names_sep, inner)