Closes https://github.com/tidyverse/tidyr/issues/1426

I think the only reason to use `glue()` was the handling in the case of empty input. This is already covered in tests (`names_sep` works with empty elements (#1185)`).

``` r
library(tidyr)

df <- tibble(
  id = 1:10e3,
  x = list(1:3)
)

bench::mark(
  unnest_wider = unnest_wider(df, x, names_sep = "_")
)

# Before
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 × 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unnest_wider    1.79s    1.79s     0.558    3.66MB     22.3

# After
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 × 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unnest_wider    207ms    227ms      4.45    3.66MB     20.8
```

<sup>Created on 2022-11-23 with [reprex v2.0.2](https://reprex.tidyverse.org)</sup>