@DavisVaughan 2022-11-23
Would you mind trying `tidyr:::vec_paste0()`? Is that too slow?