@mgirlich 2022-11-23
Much better than before but still quite expensive

```
# A tibble: 1 × 6
 expression        min   median `itr/sec` mem_alloc `gc/sec`
 <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
1 unnest_wider    288ms    318ms      3.15    6.18MB     17.3
```